chore: tests

+++ backend/src/services/post.js
@@ -1,40 +1,42 @@
-import { Post } from "../db/models/post.js";
-
-export async function createPost({ title, author, contents, tags }) {
-  const post = new Post({ title, author, contents, tags });
-  return await post.save();
-}
+import { Post } from '../db/models/post.js'
+import { User } from '../db/models/user.js'
 
 export async function listPosts(
   query = {},
-  { sortBy = "createdAt", sortOrder = "descending" } = {}
+  { sortBy = 'createdAt', sortOrder = 'descending' } = {},
 ) {
-  return await Post.find(query).sort({ [sortBy]: sortOrder });
+  return await Post.find(query).sort({ [sortBy]: sortOrder })
 }
 export async function listAllPosts(options) {
-  return await listPosts({}, options);
+  return await listPosts({}, options)
 }
 
-export async function listPostsByAuthor(author, options) {
-  return await listPosts({ author }, options);
+export async function listPostsByAuthor(authorUsername, options) {
+  const user = await User.findOne({ username: authorUsername })
+  return await listPosts({ author: user._id }, options)
 }
 
 export async function listPostByTag(tags, options) {
-  return await listPosts({ tags }, options);
+  return await listPosts({ tags }, options)
 }
 
 export async function getPostById(postId) {
-  return await Post.findById(postId);
+  return await Post.findById(postId)
 }
 
-export async function updatePost(postId, { title, author, contents, tags }) {
+export async function createPost(userId, { title, contents, tags }) {
+  const post = new Post({ title, author: userId, contents, tags })
+  return await post.save()
+}
+
+export async function updatePost(userId, postId, { title, contents, tags }) {
   return await Post.findOneAndUpdate(
-    { _id: postId },
-    { $set: { title, author, contents, tags } },
-    { new: true }
-  );
+    { _id: postId, author: userId },
+    { $set: { title, contents, tags } },
+    { new: true },
+  )
 }
 
-export async function deletePost(postId) {
-  return await Post.deleteOne({ _id: postId });
+export async function deletePost(userId, postId) {
+  return await Post.deleteOne({ _id: postId, author: userId })
 }
